Extract RepoMapBuilder from DbReader
DbReader was 1174 lines — the largest class in the codebase.
Move repo-map logic (~280 lines) into a dedicated RepoMapBuilder:
file stats queries, entrypoint scoring, module grouping, and
summary construction.

The public API (DbReader.GetRepoMap) is unchanged; it delegates
to RepoMapBuilder internally. Shared query helpers become
internal static for cross-class reuse.
